# target/classes/openapi.yaml
openapi: 3.0.3
info:
  description: "API для управления бюджетом, категориями и транзакциями"
  title: Budget Management API
  version: 1.0.0
servers:
- description: Local development server
  url: http://localhost:8080/api
paths:
  /categories:
    get:
      operationId: getAllCategories
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                type: array
          description: Успешный ответ
      summary: Получить все категории
      x-accepts:
      - application/json
    post:
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Категория успешно создана
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Неверные данные запроса
      summary: Создать новую категорию
      x-content-type: application/json
      x-accepts:
      - application/json
  /categories/{id}:
    delete:
      operationId: deleteCategory
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Категория успешно удалена
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Категория не найдена
      summary: Удалить категорию
      x-accepts:
      - application/json
    get:
      operationId: getCategoryById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Успешный ответ
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Категория не найдена
      summary: Получить категорию по ID
      x-accepts:
      - application/json
    put:
      operationId: updateCategory
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Категория успешно обновлена
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Неверные данные запроса
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Категория не найдена
      summary: Обновить категорию
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
    Category:
      example:
        color: color
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        color:
          type: string
      required:
      - color
      - id
      - name
      type: object
    CategoryRequest:
      example:
        color: color
        name: name
      properties:
        name:
          type: string
        color:
          type: string
      required:
      - color
      - name
      type: object
...
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object